update notebook server images to final v1.8.0 release
Removes the `-rc*` from all notebook images, pointing instead to the final releases.

Also fixes a bug where, previously, the vscode and rstudio images were reversed in config.yaml (the vscode-images entry had an rstudio image, and vise versa).  This may have been because the config.yaml images are in a different order than the upstream spawner_ui_config.yaml that we sync to - to avoid this in future, the config.yaml ordering was changed.