fix: expose metrics port using kubernetes_service_patch lib
This commit ensures the metrics port is exposed in the Kubernetes Service for the jupyter-controller
using the kubernetes_service_patch lib. This makes the metrics endpoint reachable from external
prometheus scraper.
This commit also changes the unit tests slightly to adapt to the added service patcher.

Part of canonical/bundle-kubeflow#564